Adding Python Code
The python code for the data analytics and visualisation of the covid data from the CSV file plus the rewriting of the data to a new CSV file. 
Included calculations and visualisations on 3 day, 5 day and 7 day moving averages as line graphs: Since 25 January 2020, since 1 January 2022 and since the data of official WA border reopening (3 March 2022).